% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmr.R
\docType{package}
\name{abmr}
\alias{abmr}
\title{Agent based modeling package}
\description{
A work-in-progress package for running agent based models in R.
}
\author{
Benjamin Gochanour \email{ben.gochanour@ou.edu}

Javi Fernandez-Lopez

Andrea Contina
}
