% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.species.R
\name{as.species}
\alias{as.species}
\title{Creates object of "species" class for input into moveSIM() and energySIM()}
\usage{
as.species(x = NA, y = NA)
}
\arguments{
\item{x}{Species origin longitude value (degrees). Required.}

\item{y}{Species origin latitude value (degrees). Required.}
}
\value{
Object of class "species" for input into moveSIM() or energySIM()
}
\description{
Here we define the geographical origin of the agents whose movement we will be modeling.
}
\examples{
myspecies <- as.species(x = -98.7, y = 34.7)
}
