% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseaseSIM_helper.R
\name{diseaseSIM_helper}
\alias{diseaseSIM_helper}
\title{Run energy-dynamic based model for one replicate}
\usage{
diseaseSIM_helper(
  sp,
  env_orig,
  env_subtract,
  days,
  sigma,
  dest_x,
  dest_y,
  mot_x,
  mot_y,
  search_radius,
  optimum_lo,
  optimum_hi,
  init_energy,
  direction,
  single_rast,
  mortality,
  energy_adj,
  disease_loc,
  disease_radius,
  disease_mortality,
  disease_energy_interact
)
}
\arguments{
\item{sp}{A species object}

\item{days}{Integer, how many days (timesteps), would you like to model}

\item{sigma}{Numeric, amount of random error}

\item{dest_x}{Numeric, destination x coordinate (longitude)}

\item{dest_y}{Numeric, destination y coordinate (latitude)}

\item{mot_x}{Numeric, movement motivation in x direction}

\item{mot_y}{Numeric, movement motivation in y direction}

\item{search_radius}{Radius of semicircle to South of current location to search for next timestep (in km)}

\item{optimum_lo}{come back}

\item{optimum_hi}{come back}

\item{init_energy}{come back}

\item{direction}{come back}

\item{mortality}{Logical, should low energy levels result in death?}

\item{disease_loc}{Dataframe of x,y coordinates specifying the location of disease clusters.}

\item{disease_radius}{Numeric (0,Infty), what is the maximum distance from the source point is capable
capable of causing disease?}

\item{disease_mortality}{Numeric (0,1] What proportion of agents who pass directly
over disease source (maximum disease load) will experience mortality? Assume mortality rate then linearly
decreases to 0 at disease_radius.}

\item{disease_energy_interact}{Numeric (0,100] Below what energy level should all agents
exposed to disease die?}

\item{env}{Raster, should represent NDVI or your environmental variable of interest}

\item{sp_poly}{Come back to this}
}
\value{
A nx3 dataset containing longitude and latitude and energy
points for all n timesteps
}
\description{
Runs agent based modeling for one replicate of a single species.
}
\keyword{internal}
