% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energySIM2.R
\name{energySIM2}
\alias{energySIM2}
\title{Runs more advanced Brownian/Ornstein–Uhlenbeck agent-based model for
multiple replicates.}
\usage{
energySIM2(
  replicates = 200,
  days = 27,
  env_rast = ndvi_raster,
  search_radius = 375,
  sigma,
  dest_x,
  dest_y,
  mot_x,
  mot_y,
  modeled_species,
  optimum_lo,
  optimum_hi,
  init_energy,
  direction = "S",
  mortality = TRUE,
  energy_adj = c(25, 20, 15, 10, 5, 0, -5, -10, -15, -20, -25),
  write_results = FALSE,
  single_rast = FALSE
)
}
\arguments{
\item{replicates}{Integer, desired number of replicates per generation}

\item{days}{Integer, How many days (timesteps), would you like to model?}

\item{env_rast}{Rasterstack or Rasterbrick with number of layers equal to n_days}

\item{search_radius}{Radius of semicircle to South of current location to search for next timestep (in km)}

\item{sigma}{Numeric, randomness parameter}

\item{dest_x}{Numeric, destination x coordinate (longitude)}

\item{dest_y}{Numeric, destination y coordinate (latitude)}

\item{mot_x}{Numeric, movement motivation in x direction}

\item{mot_y}{Numeric, movement motivation in y direction}

\item{modeled_species}{Object of class "species"}

\item{optimum_lo}{Numeric, optimal environmental value (low)}

\item{optimum_hi}{Numeric, optimal environmental value (high)}

\item{init_energy}{Numeric, initial energy in interval (0,100]}

\item{direction}{Character, mig direction, one of "N","S","E","W"}

\item{mortality}{Logical, should low energy levels result in death?}

\item{write_results}{Logical, save results to csv?}

\item{single_rast}{Logical, are you using a one-layer raster for all timesteps?}
}
\value{
#' A (days X replicates) X 7 dataframe containing data on latitude, longitude, energy,
day, agent ID, distance traveled between each timestep (in km), and change in
energy from last timestep.
}
\description{
Here, agent mortality occurs when agent reaches energy = 0 (out of 100). Agent energy
stores are dynamic, and affect search area as a multiplier, so movement speed
is directly affected by the quality of raster cells achieved. Results may be analyzed
with `energyVIZ()`. Relies on underlying function `energySIM_helper`, which is
not to be used alone.
}
\examples{
testing=energySIM(replicates=1,days=27,env_rast=ndvi_raster, search_radius=375,
sigma=.4, dest_x=-100, dest_y=25, mot_x=1, mot_y=1,
modeled_species=N_pop, my_shapefile=NOAM, optimum_lo=.4,optimum_hi=.6,
init_energy=100,direction="S",write_results=TRUE,single_rast=FALSE)
}
