% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EX1.R
\docType{data}
\name{EX1}
\alias{EX1}
\title{Example energySIM result}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(EX1)
}
\description{
This is an output for energySIM results.
}
\examples{
data(EX1)
}
\keyword{datasets}
