% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_raster.R
\docType{data}
\name{ex_raster}
\alias{ex_raster}
\title{Example Environmental Raster}
\format{
An object of class \code{RasterLayer} of dimension 200 x 200 x 1.
}
\usage{
data(ex_raster)
}
\description{
Simulated environmental raster for usage in abmR examples.
}
\examples{
data(ex_raster)
}
\keyword{datasets}
