% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generations.R
\name{generations}
\alias{generations}
\title{Runs model for multiple replicates over multiple generations}
\usage{
generations(
  n_generations = 3,
  replicates = 200,
  days = 27,
  ndvi_rast = ndvi_raster,
  my_sigma,
  modeled_species,
  my_NOAM = NOAM
)
}
\arguments{
\item{n_generations}{Integer, desired number of generations to run}

\item{replicates}{Integer, desired number of replicates per generation}

\item{days}{Integer, How many days (timesteps), would you like to model?}

\item{ndvi_rast}{COME BACK}

\item{my_sigma}{Numeric, randomness parameter}

\item{modeled_species}{Object of class "species"}

\item{my_NOAM}{COME BACK}

\item{mot_y}{Numeric, movement motivation in y direction}

\item{sp_poly}{Come back to this}

\item{current_generation}{Fed into function by function `generations`}

\item{search_radius}{Radius of semicircle to South of current location to search for next timestep (in km)}
}
\value{
A matrix
}
\description{
Runs model for multiple replicates over multiple generations
}
\examples{
Come back to this
}
