% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSIM.R
\name{genSIM}
\alias{genSIM}
\title{Runs model for multiple replicates over multiple generations}
\usage{
genSIM(
  n_generations = 3,
  replicates = 200,
  days = 27,
  ndvi_rast = ndvi_raster,
  search_radius = 375,
  my_sigma,
  dest_x,
  dest_y,
  mot_x,
  mot_y,
  modeled_species,
  my_shapefile = NOAM
)
}
\arguments{
\item{n_generations}{Integer, desired number of generations to run}

\item{replicates}{Integer, desired number of replicates per generation}

\item{days}{Integer, How many days (timesteps), would you like to model?}

\item{ndvi_rast}{COME BACK}

\item{search_radius}{Radius of semicircle to South of current location to search for next timestep (in km)}

\item{my_sigma}{Numeric, randomness parameter}

\item{dest_x}{Numeric, destination x coordinate (longitude)}

\item{dest_y}{Numeric, destination y coordinate (latitude)}

\item{mot_x}{Numeric, movement motivation in x direction}

\item{mot_y}{Numeric, movement motivation in y direction}

\item{modeled_species}{Object of class "species"}

\item{my_shapefile}{COME BACK}
}
\value{
A matrix
}
\description{
Runs model for multiple replicates over multiple generations
}
\examples{
Come back to this
}
