% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Run model for one replicate}
\usage{
move(
  sp,
  env,
  n,
  sigma,
  theta_x,
  alpha_x,
  theta_y,
  alpha_y,
  sp_poly,
  current_gen
)
}
\arguments{
\item{sp}{A species object}

\item{env}{Raster, should represent NDVI or your environmental variable of interest}

\item{n}{Integer, how many days (timesteps), would you like to model}

\item{theta_x}{Numeric, destination x coordinate (longitude)}

\item{alpha_x}{Numeric, movement motivation in x direction}

\item{theta_y}{Numeric, destination y coordinate (latitude)}

\item{alpha_y}{Numeric, movement motivation in y direction}

\item{sp_poly}{Come back to this}

\item{current_generation}{Fed into function by function `generations`}
}
\value{
A nx2 dataset containing longitude and latitude points for all n timesteps
}
\description{
Runs agent based modeling for one replicate of a single species. Used by function `generations`
to run more replicates of more species.
}
\examples{
Come back to this
}
