% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{class}
\name{species-class}
\alias{species-class}
\alias{species}
\title{A class for the input species}
\description{
A class for the input species
}
\section{Slots}{

\describe{
\item{\code{x}}{Species origin longitude value (degrees)}

\item{\code{y}}{Species origin latitude value (degrees)}

\item{\code{mass}}{Species mass}

\item{\code{wing}}{Species wing}
}}

