% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmR.R
\docType{package}
\name{abmR}
\alias{abmR}
\title{Agent based modeling package}
\description{
A work-in-progress package for running agent based models in R.
}
\author{
Benjamin Gochanour \email{ben.gochanour@gmail.com}

Javier Fernandez-Lopez \email{jflopez.bio@gmail.com}

Andrea Contina \email{acontina@usgs.gov; acontina@gmail.com}
}
