% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyVIZ.R
\name{energyVIZ}
\alias{energyVIZ}
\title{Creates a plot of energySIM() results}
\usage{
energyVIZ(
  data,
  title = "EnergySIM results",
  type = "plot",
  aspect_ratio = 4/3,
  label = FALSE,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{Data to be plotted, this object should be the output from
energySIM().}

\item{title}{Title for the plot that is output.}

\item{type}{String from "plot", "gradient", "summary_table", or "strat_table"?}

\item{aspect_ratio}{Aspect ratio, defaults to 4/3}

\item{label}{Logical, label the origin and specified final destination?}

\item{xlim}{Optionally specify desired x limits as a numeric vector: c(low,hi)}

\item{ylim}{Optionally specify desired y limits as a numeric vector: c(low,hi)}
}
\value{
A plot showing model output compared to a dashed line that represents straight line
movement from the starting point to the final destination.
}
\description{
Compares results with straight line path between origin and destination.
}
\examples{
1. Run energySIM()

EX1=energySIM(replicates=5,days=27,env_rast=ndvi_raster, search_radius=400,
sigma=.1, dest_x=-108.6, dest_y=26.2, mot_x=.9, mot_y=.9,
modeled_species=pabu.pop,
optimum_lo=.6,optimum_hi=.8,init_energy=100,
direction="S",write_results=FALSE,single_rast=FALSE,mortality = TRUE)

2. Run energyVIZ() on your result
energyVIZ(EX1,title="Visualizing EnergySIM results",type="plot", aspect_ratio=5/3,
label=TRUE)
}
