% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyVIZ.R
\name{energyVIZ}
\alias{energyVIZ}
\title{Creates a plot of energySIM() results}
\usage{
energyVIZ(data, title = "Visualizing EnergySIM results", type = "plot")
}
\arguments{
\item{data}{Data to be plotted, this object should be the output from
energySIM().}

\item{title}{Title for the plot that is output.}

\item{type}{Plot, gradient, or table?}
}
\value{
A plot showing model output compared to a dashed line that represents straight line
movement from the starting point to the final destination.
}
\description{
Compares results with straight line path (null model)
}
\examples{

}
