% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveSIM.R
\name{moveSIM}
\alias{moveSIM}
\title{Runs basic Brownianb / Ornstein Uhlenbeck agent-based model for multiple replicates.}
\usage{
moveSIM(
  replicates = 200,
  days,
  env_rast = ndvi_raster,
  search_radius = 375,
  sigma,
  dest_x,
  dest_y,
  mot_x,
  mot_y,
  modeled_species,
  optimum,
  n_failures = 4,
  fail_thresh = 0.5,
  direction = "S",
  mortality = TRUE,
  write_results = FALSE,
  single_rast = FALSE
)
}
\arguments{
\item{replicates}{Integer, desired number of replicates per run. Default 200.}

\item{days}{Integer, How many days (timesteps) would you like to model? Range (1,nlayers(env_rast))}

\item{env_rast}{Rasterstack or Rasterbrick with number of layers >= days}

\item{search_radius}{Radius of semicircle search regions (in km). Default 375.}

\item{sigma}{Numeric, randomness parameter, range (-Infty, Infty). Default 0.5.}

\item{dest_x}{Numeric, destination x coordinate (longitude)}

\item{dest_y}{Numeric, destination y coordinate (latitude)}

\item{mot_x}{Numeric, movement motivation in x direction, range (0,1], default 1.}

\item{mot_y}{Numeric, movement motivation in y direction, range (0,1], default 1.}

\item{modeled_species}{Object of class "species"}

\item{optimum}{Numeric, optimal environmental value}

\item{n_failures}{How many failures are allowable before agent experiences death (at n_failures+1). What constitutes
a failure is determined by fail_thresh, range (1,days]. Default 4.}

\item{fail_thresh}{What percentage deviation from optimum leads to death? E.g. default of 
.50 means 50 percent or greater deviation from optimum on a particular step constitutes failure.}

\item{direction}{Character, movement direction, one of "N","S","E","W". Default "S".}

\item{mortality}{Logical, should low energy levels result in death? Default T.}

\item{write_results}{Logical, save results to csv? Default F.}

\item{single_rast}{Logical, are you using a one-layer raster for all timesteps? Default F.}
}
\value{
Under "results", a (days+1 X replicates) X 7 dataframe containing data on agent_id, day, longitude, latitude,
current agent status (Alive, Stopped, or Died), distance traveled from last timestep (in km), and plot_ignore, which can be ignored by the user.
Using tidy_results() hides this column and provides overall nicer display of results.

Under "run_params", a record of function parameters used as well as missing_pct
and mortality_pct. missing_pct corresponds to the percent of rows in the results dataframe
missing information on lon/lat, which occurs when the agent has "died" or "stopped". mortality_pct
refers to the percentage of agents in the run that died.
}
\description{
Here, agent mortality occurs when agent fails to achieve suitable raster values
at least 5 days (timesteps) in a row. Agent energy stores are not dynamic, so movement
speed isn't directly affected by quality of raster cells achieved. Results may be analyzed
with `moveVIZ()`. Relies on underlying function `moveSIM_helper`, which is not to be used
alone.
}
\details{
Arguments mortality, n_failures, and fail_thresh interact with each other. If 
mortality = F, values for n_failures and fail_thresh are ignored. If mortality=T, fail_thresh
determines what consitutes a failure, and n_failures indicates how many failures are allowed before
death. Note: If n_failures=days, this is equivalent to mortality=F.
}
\examples{
# Define species object
pabu.pop = as.species(x=-98.7, y=34.7,
morphpar1=15, morphpar1mean=16, morphpar1sd=2,morphpar1sign="Pos",
morphpar2=19,morphpar2mean=18,morphpar2sd=1,morphpar2sign="Pos")

# Run function
EX2=moveSIM(replicates=5,days=27,env_rast=ndvi_raster, search_radius=550,
sigma=.1, dest_x=-108.6, dest_y=26.2, mot_x=.8, mot_y=.8,modeled_species=pabu.pop,optimum=.6, n_failures=5, fail_thresh=.40,
direction="S",write_results=TRUE,single_rast=FALSE,mortality = T)
 
# View Results in Clean Format
tidy_results(EX2,type="results")
tidy_results(EX2,type="run_params")

}
