% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveSIM.R
\name{moveSIM}
\alias{moveSIM}
\title{Runs basic Brownian/Ornstein–Uhlenbeck agent-based model for multiple replicates.}
\usage{
moveSIM(
  replicates = 200,
  days,
  env_rast = ndvi_raster,
  search_radius = 375,
  sigma,
  dest_x,
  dest_y,
  mot_x,
  mot_y,
  modeled_species,
  optimum,
  n_failures = 4,
  fail_thresh = 0.5,
  direction = "S",
  mortality = TRUE,
  write_results = FALSE,
  single_rast = FALSE
)
}
\arguments{
\item{replicates}{Integer, desired number of replicates per run. Default 200.}

\item{days}{Integer, how many days (timesteps), would you like to model?}

\item{env_rast}{Rasterstack or Rasterbrick with number of layers equal to days}

\item{search_radius}{Radius of semicircle search regions (in km). Default 375.}

\item{sigma}{Numeric, randomness parameter, range (-Infty, Infty). Default 0.5.}

\item{dest_x}{Numeric, destination x coordinate (longitude)}

\item{dest_y}{Numeric, destination y coordinate (latitude)}

\item{mot_x}{Numeric, movement motivation in x direction, range (0,1], default 1.}

\item{mot_y}{Numeric, movement motivation in y direction,range (0,1], default 1.}

\item{modeled_species}{Object of class "species"}

\item{optimum}{Numeric, optimal environmental value}

\item{n_failures}{How many failures are allowable before agent experiences death (at n_failures+1)}

\item{fail_thresh}{What percentage deviation from optimum leads to death? E.g. .50 = 50 percent.}

\item{direction}{Character, movement direction, one of "N","S","E","W". Default "S".}

\item{mortality}{Logical, should low energy levels result in death? Default T.}

\item{write_results}{Logical, save results to csv? Default F.}

\item{single_rast}{Logical, are you using a one-layer raster for all timesteps? Default F.}
}
\value{
A (days X replicates) X 5 dataframe containing data on latitude, longitude,
day, agent ID, and distance traveled between each timestep (in km).
}
\description{
Here, agent mortality occurs when agent fails to achieve suitable raster values
at least 5 days (timesteps) in a row. Agent energy stores are not dynamic, so movement
speed isn't directly affected by quality of raster cells achieved. Results may be analyzed
with `moveVIZ()`. Relies on underlying function `moveSIM_helper`, which is not to be used
alone.
}
\examples{
testing=moveSIM(replicates=1,days=27,env_rast=ndvi_raster, search_radius=375,
sigma=.4, dest_x=-100, dest_y=25, mot_x=1, mot_y=1,modeled_species=N_pop, optimum=.5,direction="S",write_results=TRUE,single_rast=FALSE)
}
