% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveSIM.R
\name{moveSIM}
\alias{moveSIM}
\title{Runs model for multiple replicates over multiple generations}
\usage{
moveSIM(
  replicates = 200,
  days = 27,
  env_rast = ndvi_raster,
  search_radius = 375,
  my_sigma,
  dest_x,
  dest_y,
  mot_x,
  mot_y,
  modeled_species,
  my_shapefile = NOAM,
  optimum
)
}
\arguments{
\item{replicates}{Integer, desired number of replicates per generation}

\item{days}{Integer, How many days (timesteps), would you like to model?}

\item{env_rast}{Rasterstack or Rasterbrick with number of layers equal to n_days}

\item{search_radius}{Radius of semicircle to South of current location to search for next timestep (in km)}

\item{my_sigma}{Numeric, randomness parameter}

\item{dest_x}{Numeric, destination x coordinate (longitude)}

\item{dest_y}{Numeric, destination y coordinate (latitude)}

\item{mot_x}{Numeric, movement motivation in x direction}

\item{mot_y}{Numeric, movement motivation in y direction}

\item{modeled_species}{Object of class "species"}

\item{my_shapefile}{COME BACK}
}
\value{
1. species_full: A (days x replicates X n_generations) X 5 dataframe containing the columns
lon, lat, day, gen, and bird_id (e.g. 1_2 means second bird of first gen.)

2. last_gen: species_full subsetted to last gen.

3. shapefiles: A list of shapefiles produced, one for each generation
}
\description{
Runs model for multiple replicates over multiple generations
}
\examples{
my_results=genSIM(n_generations = 2,replicates = 3,my_sigma = 0.7,env_rast = ndvi_raster,
modeled_species = N_pop, dest_x = -100,dest_y = 25,my_shapefile = NOAM,
mot_x = 0.9,mot_y = 0.8,days = 27)
my_results$last_gen
my_results$species_full
my_results$shapefiles
}
