% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveSIM_helper.R
\name{moveSIM_helper}
\alias{moveSIM_helper}
\title{Run model for one replicate}
\usage{
moveSIM_helper(
  sp,
  env,
  days,
  sigma,
  dest_x,
  dest_y,
  mot_x,
  mot_y,
  sp_poly,
  search_radius,
  optimum,
  direction,
  single_rast
)
}
\arguments{
\item{sp}{A species object}

\item{env}{Raster, should represent NDVI or your environmental variable of interest}

\item{days}{Integer, how many days (timesteps), would you like to model}

\item{sigma}{Numeric, amount of random error}

\item{dest_x}{Numeric, destination x coordinate (longitude)}

\item{dest_y}{Numeric, destination y coordinate (latitude)}

\item{mot_x}{Numeric, movement motivation in x direction}

\item{mot_y}{Numeric, movement motivation in y direction}

\item{sp_poly}{Come back to this}

\item{search_radius}{Radius of semicircle to South of current location to search for next timestep (in km)}

\item{current_gen}{Fed into function by function genSIM; if using this function
alone use 1}
}
\value{
A nx2 dataset containing longitude and latitude points for all n timesteps
}
\description{
Runs agent based modeling for one replicate of a single species. Used by function `generations`
to run more replicates of more species.
}
\examples{
my_results = moveSIM(sp = wiwa.pop, env = ndvi_raster, n = 27, sigma = 0.6,
dest_x = -100, dest_y = 25, mot_x = 0.9, mot_y = 0.9, search_radius = 200, current_gen = 1)
}
