% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveVIZ.R
\name{moveVIZ}
\alias{moveVIZ}
\title{Creates a plot of energySIM() results}
\usage{
moveVIZ(data, title = "Visualizing MoveSIM results")
}
\arguments{
\item{data}{Data to be plotted, this object should be the output from
moveSIM().}

\item{title}{Title for the plot that is output.}
}
\value{
A plot showing model output compared to a dashed line that represents straight line
movement from the starting point to the final destination.
}
\description{
Compares results with straight line path (null model)
}
\examples{

}
