% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveVIZ.R
\name{moveVIZ}
\alias{moveVIZ}
\title{Creates a plot of energySIM() results}
\usage{
moveVIZ(data, title = "MoveSIM results")
}
\arguments{
\item{data}{Data to be plotted, this object should be the output from
moveSIM().}

\item{title}{Title for the plot that is output.}
}
\value{
A plot showing model output compared to a dashed line that represents straight line
movement from the starting point to the final destination.
}
\description{
Compares results with straight line path (null model)
}
\examples{

1. Run moveSIM()

my_result=moveSIM(replicates=1,days=27,env_rast=ndvi_raster, search_radius=375,
sigma=.4, dest_x=-100, dest_y=25, mot_x=1, mot_y=1,modeled_species=N_pop,
my_shapefile=NOAM,optimum=.5,direction="S",write_results=TRUE,single_rast=FALSE)

2. Run energySIM() on your result
moveVIZ(my_result,title="Visualizing MoveSIM results",type="plot",aspect_ratio=4/3,
label=TRUE)

}
