% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveVIZ.R
\name{moveVIZ}
\alias{moveVIZ}
\title{Creates a plot of energySIM() results}
\usage{
moveVIZ(data, type = "plot", title = "MoveSIM results")
}
\arguments{
\item{data}{Data to be plotted, this object should be the output from
moveSIM().}

\item{type}{"plot" or "summary_table", default "plot".}

\item{title}{Title for the plot that is output.}
}
\value{
A plot showing model output compared to a dashed line that represents straight line
movement from the starting point to the final destination.
}
\description{
Compares results with straight line path between origin and destination.
}
\examples{

1. Run moveSIM()

EX2=moveSIM(replicates=5,days=27,env_rast=ndvi_raster, search_radius=550,
sigma=.1, dest_x=-108.6, dest_y=26.2, mot_x=.8, mot_y=.8,modeled_species=pabu.pop,optimum=.6, n_failures=5, fail_thresh=.40,
 direction="S",write_results=TRUE,single_rast=FALSE,mortality = T)

2. Run energySIM() on your result
moveVIZ(EX2,title="Visualizing MoveSIM results",type="plot",aspect_ratio=4/3,
label=TRUE)

}
